Move command struct and parser to command.rs
